Align symbols horizontally in rofi's presentation
This patch makes the symbols horizontally aligned in rofi's lines. I
believe this is much more easy on the eye, making a better UX.

rofi does not seem to support this natively, that is, providing a
delimiter to create columns like a table. Instead, we use the `column`
tool to make them with spaces. However, rofi prints the selected row as
is (with contiguous spaces to create the columns), so it's hard to find
the selected symbol in the output.

The solution is to make the `special_character_list` file a database
where each entry is a tab-separated row (description and symbol), and
have it compiled (by `compile.sh`) to a file of rofi lines
(`special_character_list.rofi`) and a file of symbols only
(`special_character_list.symbols`). Then we use rofi's `-format d`
option to make it return the selected row number and pick the
corresponding symbol in `special_character_list.symbols`. `wc --lines`
should return the same value for both files.

I'm using the tab character instead of `:` as the delimiter because
there can be descritions or symbols with the (common) `:` character. For
example, there was this one:

    kaomoji_:3_blushing         : (*´ω｀*)

which would type `3_blushing` with xdotool because `3_blushing` is
field 2 when using `:` as a delimiter. The tab character is not likely
to be part of any symbol so it seems safe enough.

`compile.sh` creates a file `.compiled-sha1` which contains the SHA-1
hash of the `special_character_list` file. `charpickerprint.sh` checks
that the file exists, and if it does, that its content matches the
current SHA-1 hash of the file. Otherwise, it runs `compile.sh`. So
there's no need to run `compile.sh` manually after updating the
database.

I'm also splitting the scripts into `charpickerprint.sh` which only
prints the selected symbol, if any, and `charpicker.sh` which runs
`charpickerprint.sh` and uses xdotool. This is more flexible and
backward compatible.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>