Pick user special characters if available
This patch adds support for user symbols. Those symbols are added to the
project's provided symbol list, as explained in `README.md`.

The purpose of this feature is for a user to have its own symbols which
are not necessarily interesting to the general public while still being
able to update the official project's list.

The new script, `effective_special_character_list.sh`, outputs the
concatenation of the provided list of symbols (`special_character_list`)
and the user list of symbols
(`~/.config/charpicker/special_character_list`), if any. `compile.sh`
and `charpickerprint.sh` use this script instead of reading from
`special_character_list` directly.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>